Finished game logic and debug tests.
i.e. The game now functions in accordance with the dump file provided on
the course website. I had to delete the java RNG from the start and
reimplement it in javascript using an arbitrary precision arithmetic
library. This was due to the fact that node can seemingly only interface
with child processes in an asynchronous fashion, which is really cool for
most applications, but all of the game code I've written is synchronous.
Much refactoring would have to be done to get it asynchronous which would
probably cause a bit of function despatch overhead and make for some
unintuitive code.

I'm sure the code is a bit messy now too, since yesterday was a massive
hackfest; piling mountains of highly interdependant code together and
hoping for the best. Now that the behavioural test is set up and passing,
I can go through it all and refactor before transitioning to work on the GUI and such.