updated build system and readme. Removed linting stuff.
Linting lots of separate files which depend on each other means that
things which are not defined in a particular file throw errors when they
will be defined once the program is all linked together. While there are
some benefits to linting, I can count on one hand the number of times
where jshint has detected a syntax error that I needed to fix. And these
would have been easily caught and fixed by running the program in a
browser. 99.999% of
the time, linting only exposes formatting errors, and the undefined variable
errors described above. Formatting errors are not really errors so much as
frowned-upon levels of indentation and whatnot. Undefined variable errors
are unavoidable without writing lots of extraneous statements like the
following:
	some_var = some_var || {};
It is for these reasons that I've abandoned linting. So long, old friend.